microservices/cast-service/requirements.txt
asyncpg
databases[postgresql]
fastapi
SQLAlchemy
psycopg2
fastapi-cache2[redis]